+++ CHANGELOG.md
@@ -4,7 +4,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
 ### Changed
-- `StringValue`: Switch hex conversion to use `IntOutput`
+- `StringValue`: switch hex conversion to use `IntOutput`
+- `pom.xml`: update Kotlin version to 1.7.21
 
 ## [0.88] - 2023-02-14
 ### Changed